코틀린 테스트코드 작성 추가

+++ carecat/src/main/kotlin/com/foameraserblue/carecat/web/repository/marker/MarkerRepositoryCustom.kt
@@ -3,5 +3,5 @@ package com.foameraserblue.carecat.web.repository.marker
 import com.foameraserblue.carecat.web.dto.marker.AdjMarkerResponseDto
 
 interface MarkerRepositoryCustom {
-    fun adjMarker(lat: Double, lng: Double, distance: Int): List<AdjMarkerResponseDto>
+    fun adjMarker(lat: Double, lng: Double, distance: Double): List<AdjMarkerResponseDto>
 }
